Fix tests.
